bus/pyproject.toml
[tool.poetry]
name = "bus"
version = "0.1.0"
description = ""
authors = ["intey <churin.d.r@sberbank.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
